time-tracker/mac-app/requirements.txt
# Install with: /Users/z/miniforge3/bin/pip install -r requirements.txt
# Or: pip3 install rumps requests python-dotenv

rumps>=0.4.0
requests>=2.28.0
python-dotenv>=1.0.0
